rename: types -> CategoryChart변경 및 이동
